Use existing record
Update launch_restore_request to get or create a record, if it doesn't exist. If a new record doesn't have a user_data record that has an 'aws_s3_path' key, there will be no attempt to restore the object.
This is needed to support re-trying restore requests in case they've timed out. Next step is to wire this into the DAG, to retry the download request.