upgrade merge requirements
The one line command left duplicates, and disregarded versions. When I wanted to use a pre-release of an update of bdrc-bag, `bdrc-docker.sh` emitted incorrect version (just pulled the last released, 0.0.7)

By using a python script, we can now pull one version of each library, and the latest, or the widest range. This allows for bdrc-sync requirements and airflow requirements to be successfully merged.