modulate feeding see note
the feeder dag slows down the sync dag unnecessarily. By trying to feed 20 archives, this slows down the feeder for hours. This attempts to tie the feeder count to the number of instances, so is the minimal number needed to keep sync dag running. At worst, one sync dag will be empty while the last feeder completes.